temporary/WinterSchoolExamples/requirements.txt
moviepy==2.1.1
envisionhgdetector==0.0.5.0
jupyter
librosa
scipy
plotly
nbformat>=4.2.0
shapedtw
umap-learn
numpy
networkx
matplotlib
scikit-image
plotly==5.5.0
dash-bootstrap-components